Выносит функции в util.js
